pyproject.toml
[project]
name = "wm-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12.2",
    "colorama>=0.4.6",
    "httpx>=0.28.1",
    "mcp[cli]>=1.6.0",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "scikit-learn>=1.3.0",
]
